Write stories/blogs on any 2  projects in 2000 words.
With your name as an author which will be then featured on your resume.

The article should contain the following sub-topics

1. Problem Definition.

2. Data Analysis.

3. EDA Concluding Remark.

4. Pre-Processing Pipeline.

5. Building Machine Learning Models.

6. Concluding Remarks.


Please note that you can choose any two topics of your choice. Please be clear on what your problem statement is, what your analysis and predictions are, and what are your final findings.